feat: manifestData write to document
